lowband_cam/requirements.txt
requests
pathlib
boto3
pillow